BICsCom

Documentation for BICsCom.

BICsCom.CylinderCache

BICsCom.SamplingPoints

BICsCom.Square

BICsCom.build_cylinder_cache

BICsCom.cartesian2polar

BICsCom.get_coeff

BICsCom.get_samplingpoints

BICsCom.CylinderCache — Type

CylinderCache

Cache for Bessel functions and Neumann function required in the spherical expansion outside the cylinder.

Fields

r: the radius of the cylinder

orders: the orders of Bessel functions and Neumann functions

J_inn: Bessel functions inside the cylinder of order in orders

J_ext: Bessel functions outside the cylinder of order in orders

Y_ext: Neumann functions outside the cylinder of order in orders

BICsCom.SamplingPoints — Type

SamplingPoints

The sampling points along the boundary of a square. The points are placed along the boundary counter clockwisely (bottom –> left –> top –> right).

Fields

n::Int64: the number of the sampling points on one edge

cc::Matrix{Float64}: the Cartesian coordinates of all sampling points

pc::Matrix{Float64}: the Polar coordinates of all sampling points

ν::Matrix{Float64}: the outer normal derivative of all sampling points

BICsCom.Square — Type

Square

A square in $\mathbb{R}^2$. The boundary of the square is the boundary of a periodic cell. We establish a Dirichlet-to-Neumann (DtN) map on the boundary based on the spherical expansion of a solution to the Helmholtz equation.

Fields

center::Vector{Float64}: the center of the square

semi::Float64: the semi length of the edge

BICsCom.build_cylinder_cache — Function

build_cylinder_cache(Ne, k, r, inn = 1.0, ext = 1.0)

Build a CylinderCache for computing the coefficients in the spherical expansion outside the cylinder.

Arguments

Ne: the number of terms in the truncated spherical expansion

k: the wavenumber

r: the radius of the cylinder

inn: the dielectric constant inside the cylinder

ext: the dielectric constant outside the cylinder

BICsCom.cartesian2polar — Method

cartesian2polar(cc::Matrix{Float64})
cartesian2polar!(pc::Matrix{Float64}, cc::Matrix{Float64})

Change coordinates from Cartesian to Polar.

BICsCom.get_coeff — Function

get_coeff(Ne, cydc::CylinderCache, inn = 1.0, ext = 1.0)

Compute the coefficients in the spherical expansion outside the cylinder.

Arguments

Ne: the number of terms in the truncated spherical expansion

cydc: the CylinderCache

inn: the dielectric constant inside the cylinder

ext: the dielectric constant outside the cylinder

BICsCom.get_samplingpoints — Method

get_samplingpoints(sq::Square, n::Int64)

Generate the sampling points along the bounary of a square sq.

Arguments

sq: a square

n: the number of sampling points on each edge