API

Spherical expansion

Extension: CMakieExt

CMakieExt.plot_square! — Function

plot_square!(ax, sq::Square)

Plot the square sq.

CMakieExt.plot_samplingpoints! — Function

plot_samplingpoints!(ax, sp::SamplingPoints; 
                     skwargs = (marker = 'x', color = :red), 
                     akwargs = (lengthscale=0.05, color=:red, shaftwidth=1, tipwidth=4))

Plot the sampling points sp with corresponding normal vectors.

Arguments

skwargs: keyword arguments for scatter in CairoMakie

akwargs: keywork arguments for arrows2d in CairoMakie

CMakieExt.plot_min_svals! — Function

plot_min_svals!(ax, ks, msvals; 
                skwargs = (color = :tomato, marker = 'x', markersize = 5), 
                lkwargs = (color = :blue))

Plot the minimal singular values for frequencies in ks.

Arguments

ks: the range of frequencies we sweep

msvals: the minimal singular values corresponding to the frequencies in ks

skwargs: keyword arguments for scatter in CairoMakie

lkwargs: keyword arguments for lines in CairoMakie