API

BICsCom.BoundaryCache

BICsCom.CylinderCache

BICsCom.SamplingPoints

BICsCom.Square

BICsCom.apply_bc

BICsCom.apply_tbc!

BICsCom.assemble_dtn

BICsCom.assemble_tbc

BICsCom.beta_m

BICsCom.build_boundary_cache

BICsCom.build_cylinder_cache

BICsCom.cartesian2polar

BICsCom.get_coeff

BICsCom.get_samplingpoints

CMakieExt.plot_min_svals!

CMakieExt.plot_samplingpoints!

CMakieExt.plot_square!

BICsCom.BoundaryCache — Type

BoundaryCache

Fields

orders::Vector{Int64}: the orders of Bessel functions and Neumann functions

J_ext::Matrix{Float64}: Bessel functions at the sampling points of order in orders

Y_ext::Matrix{Float64}: Neumann functions at the sampling points of order in orders

BICsCom.CylinderCache — Type

CylinderCache

Cache for Bessel functions and Neumann function required in the spherical expansion outside the cylinder.

Fields

r: the radius of the cylinder

orders: the orders of Bessel functions and Neumann functions

J_inn: Bessel functions inside the cylinder of order in orders

J_ext: Bessel functions outside the cylinder of order in orders

Y_ext: Neumann functions outside the cylinder of order in orders

BICsCom.SamplingPoints — Type

SamplingPoints

The sampling points along the boundary of a square. The points are placed along the boundary counter clockwisely (bottom –> left –> top –> right).

Fields

n::Int64: the number of the sampling points on one edge

cc::Matrix{Float64}: the Cartesian coordinates of all sampling points

pc::Matrix{Float64}: the Polar coordinates of all sampling points

ν::Matrix{Float64}: the outer normal derivative of all sampling points

BICsCom.Square — Type

Square

A square in $\mathbb{R}^2$. The boundary of the square is the boundary of a periodic cell. We establish a Dirichlet-to-Neumann (DtN) map on the boundary based on the spherical expansion of a solution to the Helmholtz equation.

Fields

center::Vector{Float64}: the center of the square

semi::Float64: the semi length of the edge

BICsCom.apply_bc — Method

apply_bc(Λ, n::Int64; α::Float64 = 0.0, period::Float64 = 2π)

Impose the periodic boundary condition or quasi-period boundary condition for standing Bloch modes or propagating Bloch modes, respectively.

Arguments

Λ: the DtN matrix obtained by the spherical expansion

n: the number of sampling points on each edge

α: the Bloch wavenumber

period: the period of the periodic slab

BICsCom.apply_tbc! — Method

apply_tbc!(Δ, sp::SamplingPoints, k; α = 0.0, homo = 1.0, period = 2π)

Impose the Transparent Boundary Condition for the top and bottom boundary.

Arguments

Δ: the matrix after imposing the periodic (quasi-periodic) boundary condition, see apply_bc

sp: all sampling points

k: the wavenumber

α: the Bloch wavenumber

homo: the dielectric constant of the homogeneous medium outside the periodic slab

period: the period of the periodic slab

BICsCom.assemble_dtn — Method

assemble_dtn(sp::SamplingPoints, Ne, coeff, bc::BoundaryCache, k, ext; mode::Symbol=:te, homo=1.0)

Compute the Dirichlet-to-Neumann matrix on a square based on the spherical expansion.

Arguments

sp: the sampling points

Ne: the number of terms in the truncated spherical expansion

coeff: from get_coeff, coefficients in front of Bessel functions and Neumann functions in the spherical expansion outside the cylinder

bc: the BoundaryCache

k: the wavenumber

ext: the dielectric constant outside the cylinder

mode: the polarizations, supported types: :te, :tm

homo: the dielectric constant of the homogeneous medium outside the periodic slab

Note

The argument homo is only needed when we consider TM polarization, i.e., mode = :tm

BICsCom.assemble_tbc — Method

assemble_tbc(x, beta, fmodes; α = 0.0, period = 2π)

Construct the Transparent Boundary Condition matrix for top and bottom boundaries.

Arguments

x: x coordinates of sampling points on the top or the bottom boundary

beta: get from beta_m.

fmodes: the orders in the truncated Fourier serie

α: the Bloch wavenumber

period: the period of the periodic slab

BICsCom.beta_m — Method

beta_m(m::Int64, k; α = 0.0, homo = 1.0, period = 2π)

Compute $\beta_{m}$ given by

\[\beta_m = \sqrt{k^2 - (m + \frac{2\pi}{\text{period}})^2}.\]

Arguments

m: the order of the Fourier mode

k: the wavenumber

α: the Bloch wavenumber

homo: the dielectric constant of the homogeneous medium outside the periodic slab

period: the period of the periodic slab

BICsCom.build_boundary_cache — Function

build_boundary_cache(sp::SamplingPoints, Ne, k, ext = 1.0)

Build a BoundaryCache for computing the Dirichlet-to-Neuamnn map based on the spherical expansion.

Arguments

sp: the sampling points

Ne: the number of terms in the truncated spherical expansion

k: the wavenumber

ext: the dielectric constant outside the cylinder

Note

J_ext and Y_ext are matrices of Bessel functions and Neumann functions, respectively. Their rows are indexed by radii (the sampling points) and columns are indexed by orders

We uitilize the symmetry of the square. So this function is not applicable for general contours

BICsCom.build_cylinder_cache — Function

build_cylinder_cache(Ne, k, r, inn = 1.0, ext = 1.0)

Build a CylinderCache for computing the coefficients in the spherical expansion outside the cylinder.

Arguments

Ne: the number of terms in the truncated spherical expansion

k: the wavenumber

r: the radius of the cylinder

inn: the dielectric constant inside the cylinder

ext: the dielectric constant outside the cylinder

BICsCom.cartesian2polar — Method

cartesian2polar(cc::Matrix{Float64})
cartesian2polar!(pc::Matrix{Float64}, cc::Matrix{Float64})

Change coordinates from Cartesian to Polar.

BICsCom.get_coeff — Function

get_coeff(Ne, cydc::CylinderCache, inn = 1.0, ext = 1.0; mode::Symbol = :te)

Compute the coefficients in the spherical expansion outside the cylinder.

Arguments

Ne: the number of terms in the truncated spherical expansion

cydc: the CylinderCache

inn: the dielectric constant inside the cylinder

ext: the dielectric constant outside the cylinder

mode: the polarizations, supported types: :te, :tm

BICsCom.get_samplingpoints — Method

get_samplingpoints(sq::Square, n::Int64)

Generate the sampling points along the bounary of a square sq.

Arguments

sq: a square

n: the number of sampling points on each edge

Spherical expansion

Extension: CMakieExt

CMakieExt.plot_square! — Function

plot_square!(ax, sq::Square)

Plot the square sq.

CMakieExt.plot_samplingpoints! — Function

plot_samplingpoints!(ax, sp::SamplingPoints; 
                     skwargs = (marker = 'x', color = :red), 
                     akwargs = (lengthscale=0.05, color=:red, shaftwidth=1, tipwidth=4))

Plot the sampling points sp with corresponding normal vectors.

Arguments

skwargs: keyword arguments for scatter in CairoMakie

akwargs: keywork arguments for arrows2d in CairoMakie

CMakieExt.plot_min_svals! — Function

plot_min_svals!(ax, ks, msvals; 
                skwargs = (color = :tomato, marker = 'x', markersize = 5), 
                lkwargs = (color = :blue))

Plot the minimal singular values for frequencies in ks.

Arguments

ks: the range of frequencies we sweep

msvals: the minimal singular values corresponding to the frequencies in ks

skwargs: keyword arguments for scatter in CairoMakie

lkwargs: keyword arguments for lines in CairoMakie

BoundaryCache
Fields

CylinderCache
Cache for Bessel functions and Neumann function required in the spherical expansion outside the cylinder.

SamplingPoints
The sampling points along the boundary of a square.

Square
A square in .